requirements.txt
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
greenlet==2.0.1
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pycryptodomex==3.16.0
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
snowflake-connector-python==2.8.3
snowflake-sqlalchemy==1.4.4
SQLAlchemy==1.4.44
stack-data==0.6.2
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
xlrd==2.0.1

scipy
scikit-learn
matplotlib
seaborn
plotly
altair
bokeh
pillow
opencv-python
streamlit
